Add legend below calendar
