+++ app/views/ExposureHistory/CalendarDay.js
@@ -1,14 +1,8 @@
-import React from 'react';
 import styled from '@emotion/native';
 import dayjs from 'dayjs';
-import Color from '../../constants/colors';
+import React from 'react';
 
-const Risk = {
-  Unknown: 'unknown',
-  None: 'none',
-  Possible: 'possible',
-  Today: 'today',
-};
+import { DataCircle, Risk } from './DataCircle';
 
 /**
  * Show a circular calendar day with risk level coloring.
@@ -43,11 +37,9 @@ export const CalendarDay = ({
   return (
     <DayContainer>
       {showMonthLabel && <DayOfWeek>{date.format('MMM')}</DayOfWeek>}
-      <Circle riskLevel={riskLevel}>
-        <CircleInnerText riskLevel={riskLevel}>
-          {date.format('D')}
-        </CircleInnerText>
-      </Circle>
+      <DataCircle size={36} riskLevel={riskLevel}>
+        {date.format('D')}
+      </DataCircle>
     </DayContainer>
   );
 };
@@ -55,59 +47,6 @@ export const CalendarDay = ({
 const DayContainer = styled.View`
   align-items: center;
   padding: 4px;
-`;
-
-/** Get bg color from risk level and theme warning/success */
-const getBgColor = ({ riskLevel, theme }) => {
-  if (riskLevel === Risk.Possible) {
-    return theme.warning;
-  }
-
-  if (riskLevel === Risk.None) {
-    return theme.success;
-  }
-
-  return Color.WHITE;
-};
-
-/** Get border color from theme and risk level */
-const getBorder = ({ riskLevel, theme }) => {
-  let color = 'transparent';
-
-  if (riskLevel === Risk.Unknown) {
-    color = theme.border;
-  }
-
-  if (riskLevel === Risk.Today) {
-    color = theme.textPrimaryOnBackground;
-  }
-
-  return `1px solid ${color}`;
-};
-
-const Circle = styled.View`
-  background-color: ${getBgColor};
-  border: ${getBorder};
-  width: 36px;
-  height: 36px;
-  border-radius: 18px;
-  align-items: center;
-  align-content: center;
-  flex-direction: row;
-`;
-
-const TEXT_COLOR_MAP = {
-  [Risk.None]: '#fff',
-  [Risk.Possible]: '#fff',
-  [Risk.Unknown]: '#dadada',
-};
-
-const CircleInnerText = styled.Text`
-  color: ${({ riskLevel }) => TEXT_COLOR_MAP[riskLevel]};
-  text-align: center;
-  flex: 1;
-  font-size: 12px;
-  font-weight: bold;
 `;
 
 export const DayOfWeek = styled.Text`
